virt-api: Extract SEV handler in its own file
